Removes reducer, renames action.js.
