--- project/src/store/reducer.js
@@ -1,76 +1,0 @@
-import {DEFAULT_CITY, SortTypes, AuthorizationStatus} from '../const';
-import {ActionType} from './action';
-
-const initialState = {
-  activeCity: DEFAULT_CITY,
-  offers: [],
-  room: {},
-  reviews: [],
-  offersNearby: [],
-  userData: {},
-  sortType: SortTypes.POPULAR,
-  authorizationStatus: AuthorizationStatus.UNKNOWN,
-  isDataLoaded: false,
-  isRoomDataLoaded: false,
-};
-
-export const reducer = (state = initialState, action) => {
-  switch (action.type) {
-    case ActionType.LOAD_OFFERS:
-      return {
-        ...state,
-        offers: action.payload,
-        isDataLoaded: true,
-      };
-    case ActionType.LOAD_ROOM:
-      return {
-        ...state,
-        room: action.payload,
-        isRoomDataLoaded: true,
-      };
-    case ActionType.LOAD_REVIEWS:
-      return {
-        ...state,
-        reviews: action.payload,
-        isDataLoaded: true,
-      };
-    case ActionType.LOAD_OFFERS_NEARBY:
-      return {
-        ...state,
-        offersNearby: action.payload,
-      };
-    case ActionType.REQUIRED_AUTHORIZATION:
-      return {
-        ...state,
-        authorizationStatus: action.payload,
-      };
-    case ActionType.LOAD_USER_DATA:
-      return {
-        ...state,
-        userData: action.payload,
-      };
-    case ActionType.LOGOUT:
-      return {
-        ...state,
-        authorizationStatus: AuthorizationStatus.NO_AUTH,
-      };
-    case ActionType.CHANGE_CITY:
-      return {
-        ...state,
-        activeCity: action.payload,
-        isDataLoaded: true,
-      };
-    case ActionType.CHANGE_SORT_TYPE:
-      return {
-        ...state,
-        sortType: action.payload,
-      };
-    case ActionType.SET_IS_ROOM_DATA_LOADED:
-      return {
-        ...state,
-        isRoomLoaded: action.payload,
-      };
-    default:
-      return state;
-  }
-};
